Fix WebKit load failure. Add (untested) support for AWS hosting...
Also, remove some onused cruft.

The WebKit load failure had to do with how 'content.js' was being
imported. Removing the `export-loader` settings for 'content.js'
and loading it with a raw `require` statement enables
`SQUIDSPACE.addAutoloadModule()` to execute in global scope as
expected by `SQUIDSPACE`. There is no need for a `content` symbol
to be exported or even exist.